Fixed rare bug
When the user was not logged in and used the manufacturing tool, there
was a rare case when he would trigger an exception.

To trigger the exception the user had to send the job parameter form to
get to the result page. If he returned to the job parameter form by
using the breadcrumbs for example and then logged in, he would trigger
the exception.

That's fixed now.